+++ src/main/java/com/dute/officialNetwork/domain/repository/ProductCaseStructureRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dute.officialNetwork.domain.entity.ProductCaseStructure;
 
-public interface ProductCaseStructureRepository extends JpaRepository<ProductCaseStructure, Long> {
+public interface ProductCaseStructureRepository extends JpaRepository<ProductCaseStructure, Integer> {
 }
